requirements.txt
alembic==1.0.11
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
numpy==1.16.4
oauth2client==4.1.3
pandas==0.24.2
pdfkit==0.6.1
pendulum==2.0.5
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.8.0
python-editor==1.0.4
python-gflags==3.1.2
pytz==2019.1
pytzdata==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.5
urllib3==1.25.3
Werkzeug==0.15.4
WTForms==2.2.1
XlsxWriter==1.1.8
